Improves RAG server architecture and transport.
