+++ package.json
@@ -9,6 +9,7 @@
     "dev": "tsc --watch",
     "inspector": "npx @modelcontextprotocol/inspector node build/index.js",
     "inspector:http": "powershell -NoProfile -Command \"$env:ENABLE_HTTP_MCP_TRANSPORT=1; npx @modelcontextprotocol/inspector http://localhost:3000/mcp --transport http\"",
+    "clean": "rimraf build",
     "lint": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
     "format": "prettier . --write",
@@ -30,6 +31,7 @@
     "eslint": "^9.35.0",
     "eslint-config-prettier": "^10.1.8",
     "prettier": "^3.6.2",
+    "rimraf": "^5.0.10",
     "typescript": "^5.9.2",
     "typescript-eslint": "^8.42.0"
   }
